+++ src/tunemeta/__init__.py
@@ -1,4 +1,4 @@
 from tunemeta.models import TrackMetadata
 
 __all__ = ["TrackMetadata"]
-__version__ = "1.0.1"
+__version__ = "1.0.2"